Day 32: Learned JSON handling and built Student Database System

+++ Day32_JSON_Handling/D32_Practice2_Employee_Record_Manager/main.py
@@ -43,7 +43,7 @@ while True:
             "emp_id" : int(input("\nEnter Employee ID:")),
             "name" : input("Enter Employee Name:"),
             "dept" : input("Enter Department:"),
-            "salary" : float(input("Enter Employee Salary:"))
+            "salary" : int(input("Enter Employee Salary:"))
     }
